+++ src/domain/services/controllers/CheckpointController.js
@@ -86,6 +86,8 @@ export default class CheckpointController {
 
       /** @type {CorePersistence} */
       const persistence = h._persistence;
+      /** @type {import('../../../ports/CheckpointStorePort.js').default|null} */
+      const checkpointStore = /** @type {import('../../../ports/CheckpointStorePort.js').default|null} */ (h._checkpointStore);
       const checkpointSha = await createCheckpointCommit({
         persistence,
         graphName: h._graphName,
@@ -96,6 +98,7 @@ export default class CheckpointController {
         crypto: h._crypto,
         codec: h._codec,
         ...(indexTree ? { indexTree } : {}),
+        ...(checkpointStore ? { checkpointStore } : {}),
       });
 
       const checkpointRef = buildCheckpointRef(h._graphName);
@@ -158,7 +161,9 @@ export default class CheckpointController {
     }
 
     try {
-      return await loadCheckpoint(h._persistence, checkpointSha, { codec: h._codec });
+      /** @type {import('../../../ports/CheckpointStorePort.js').default|null} */
+      const checkpointStore = /** @type {import('../../../ports/CheckpointStorePort.js').default|null} */ (h._checkpointStore);
+      return await loadCheckpoint(h._persistence, checkpointSha, { codec: h._codec, ...(checkpointStore ? { checkpointStore } : {}) });
     } catch (err) {
       const msg = err instanceof Error ? err.message : '';
       if (